[add] 検索バーを追加
